+++ src/deepsparse/server/server.py
@@ -166,16 +166,6 @@ class Server:
         def _info():
             return self.server_config
 
-        @app.get("/ping", tags=["health"], response_model=CheckReady)
-        @app.get("/v2/health/ready", tags=["health"], response_model=CheckReady)
-        @app.get("/v2/health/live", tags=["health"], response_model=CheckReady)
-        def _check_health():
-            return CheckReady(status="OK")
-
-        @app.get("/v2", tags=["metadata", "server"], response_model=str)
-        def _get_server_info():
-            return "This is the deepsparse server. Hello!"
-
         return app
 
     def _set_pytorch_num_threads(self):
